Tests+LibSSH/SFTP: Share SFTP::Attributes definition
